refactor: move calculation logic to utils for better maintainability
